Create module for loading tasks
